# openapi/device-and-scenario-service-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Device & Scenario API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {

    "/devices": {
      "get": {
        "summary": "Get user devices",
        "description": "Получить все устройства пользователя",
        "parameters": [
          {
            "name": "homeId",
            "in": "query",
            "description": "Фильтр по дому",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of devices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "name": { "type": "string" },
                      "type": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "name": { "type": "string" }
                        }
                      },
                      "status": {
                        "type": "string",
                        "enum": ["online", "offline"]
                      },
                      "scenarios": {
                        "type": "array",
                        "items": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    
    "/devices/{deviceId}": {
      "get": {
        "summary": "Get device by ID",
        "description": "Получить конкретное устройство",
        "parameters": [
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Device details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "name": { "type": "string" },
                    "type": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "name": { "type": "string" }
                      }
                    },
                    "status": {
                      "type": "string",
                      "enum": ["online", "offline"]
                    },
                    "scenarios": {
                      "type": "array",
                      "items": { "type": "integer" }
                    },
                    "lastTelemetry": {
                      "type": "object",
                      "properties": {
                        "timestamp": { "type": "integer" },
                        "metrics": { "type": "object" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Device not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },

    "/device/add": {
      "post": {
        "summary": "Add device",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "type"],
                "properties": {
                  "name": { "type": "string" },
                  "type": { 
                    "type": "integer",
                    "description": "DeviceType ID"
                  },
                  "scenarios": {
                    "type": "array",
                    "items": { "type": "integer" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device created"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },

    "/device/{deviceId}": {
      "patch": {
        "summary": "Update device",
        "parameters": [
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "type": { 
                    "type": "integer",
                    "description": "DeviceType ID"
                  },
                  "scenarios": {
                    "type": "array",
                    "items": { "type": "integer" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device updated"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Device not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },

      "delete": {
        "summary": "Delete device",
        "parameters": [
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Device deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Device not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },

    "/scenario/add": {
      "post": {
        "summary": "Add scenario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "code"],
                "properties": {
                  "name": { "type": "string" },
                  "code": { "type": "string" },
                  "triggers": { 
                    "type": "object",
                    "description": "JSON triggers"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scenario created"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },

    "/scenario/{scenarioId}": {
      "patch": {
        "summary": "Update scenario",
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "code": { "type": "string" },
                  "triggers": { 
                    "type": "object",
                    "description": "JSON triggers"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scenario updated"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Scenario not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete scenario",
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Scenario deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Scenario not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {}
}